Change ID columns to use v7 UUIDs.
The main advantage of UUIDs instead of incrementing integers is that the
IDs on different tables won't be the same, so you can't accidentally use
an ID from the wrong table by accident. v7 UUIDs are nicer than v4 since
they are ordered, but unfortunately, won't be included in PostgreSQL
until v18. For now, we have a custom implementation that can be removed
once there is native support.