Remove _param naming on function arguments
Before we used views, I believe we needed this to avoid name conflicts
with the return table definitions. Now, there's no conflict so we can
use the simpler naming.

This also removes the extra variables in the `pgledger_create_transfers`
function and uses the fields from the `transfer_request` directly for
simplicity.