Add optional JSON metadata when creating transfers
Update `pgledger_create_transfer` and `pgledger_create_transfers` to
take an optional JSON `metadata` argument, which is recorded alongside
the transfer.

Also update the example scripts to show how to pass and use the new
metadata.

For simplicity, the `transfer_requests` argument is no longer variadic
in the function version which takes all arguments. This allows us to
pass in the arguments in any order with the same syntax.